Changed Long to Integer

+++ Backend/Trello_Project/Trello_Project/src/main/java/Group_15/Trello_Project/board/repository/BoardRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BoardRepository extends JpaRepository<BoardModel, Long>
+public interface BoardRepository extends JpaRepository<BoardModel, Integer>
 {
 
 }
